ConstrainedDynamicsVis

ConstrainedDynamicsVis is a visualizer for the rigid body dynamics package ConstrainedDynamics. It is built on top of MeshCat and provides an easy to use interface.

After simulating a mechanism (see ConstrainedDynamics for details), you obtain a Storage object. Visualization can be performed with

using ConstrainedDynamicsVis

visualize(mechanism, storage)

In the visualizer, the Scene -> meshcat tab can be used to show and hide certain bodies. And the Animations can be used to restart and modify the visualization. Video recording is also possible (returns a tar-file).

ConstrainedDynamicsVis.visualize — Method

visualize(mechanism, storage; env, showframes)

Visualize a mechanism with a trajectory stored in storage.

Available kwargs

showframes: Display the coordinate frames of the bodies.

env: Choose the visualization environment ("blink", "browser", "editor").

ConstrainedDynamicsVis.convertvideo — Method

convertvideo(file; pathout, filenameout)

Convert the tar-file obtained from a simulation to an mp4-video.

Available kwargs

pathout: The output path.

filename: The new file name.